fix(build,frontend): faster retrieval of dependencies, without sudo
- Centralises node_modules folders scattered across the parts for the
frontend (frontend, frontend_lib, apps) in one place, and use symbolic links
for the node_modules of each project to the shared node_modules folder.
This allows speeding up the retrieval of node dependencies a lot.

A script builds a super package.json from all the package.json files of
the frontend folders, ensuring that the versions of the dependencies are all
compatible with each other. This file is stored in an hidden folder:
 .frontend-apps-node-modules

- Gets rid of npm link in favor a raw symbolic link, removing the need to be
root to build Tracim (except for installing the system dependencies).

Signed-off-by: Raphaël Jakse <raphael.github@jakse.fr>