fix(back, front): icons and the Travis configuration
Ref: #4169

This commit introduces a workaround to make the tests work by fetching the value of %(here)s to the environment variable TRACIM_HERE_PATH, if set, in config.py (for default values hard-coded in the backend). This is a temporary solution, a proper solution as well as a good name for this environment variable needs to be found if we decide to keep such a feature.